Add files via upload
These are all the html, css & js docs I have for the main pages of the whow website.  I have also uploaded the original images.  You will notice that all the images that are being used are directly from the web.  Feel free to change that to the original images here if possible.

Minimum Viable Product:
1.	HOMEPAGE
2.	ABOUT US
3.	SEE THE IMPACT(embedded video)
4.	CONTACT US (Form)
5.	FAQ
Nice to have:
Podcast of WHOW Members
WHOW Minutes Blog